refactor(core): address Hermes review on page_index changes
- Tighten `_build_scan_tasks` return type from bare `list` to
  `list[Coroutine[Any, Any, list[DetectedHeader]]]`; same for the
  `new_headers_tasks` local in `_detect_and_fill_gaps`.
- Clarify the `overlap_chars` docstring (who'd want to override it).
- Rewrite the semaphore init comment to state intent (extractor is
  only instantiated from inside an async entry point) rather than
  describing CPython internals.
- Move `TestMakeLmPropagatesTimeout` from `test_page_index.py` to
  `test_extraction_engine.py` so test locality matches source
  locality (`_make_lm` lives in `extraction/engine.py`).

No behavioral change. Not addressing Hermes's claim that
`asyncio.Semaphore()` in `__init__` emits a DeprecationWarning — that
was incorrect for Python 3.12; the construction is safe here because
`AsyncMarkdownPageIndex` is only instantiated from inside
`index_document`, which always runs under an event loop. Kept the
eager init; only the misleading comment was reworded.
